backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==1.4.53
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
email-validator==2.0.0
itsdangerous==2.1.2
Werkzeug==2.3.7
gunicorn==21.2.0
psycopg2-binary==2.9.7
resend==0.6.0